Refactor DecoderAdapter3D for improved pretrained VAE integration and feature handling
This commit updates the `_DecoderAdapter3D` class to conditionally initialize the `to_latent` layer based on the availability of a pretrained VAE decoder. It modifies the forward pass to handle both scenarios—using the latent representation when a pretrained decoder is present and directly processing features when it is not. Additionally, the documentation in `ARCHITECTURE.md` is enhanced to clarify the distinct roles of the FeatureProjector and DecoderAdapter, improving understanding of the model architecture.

+++ neurons/callbacks/tensorboard.py
@@ -130,9 +130,9 @@ def _render_cov_glyphs(
 
     result = []
     for b in range(B):
-        bg = rearrange(img_rgb[b].detach().cpu(), "c h w -> h w c").numpy().copy()
+        bg = rearrange(img_rgb[b].detach().cpu().float(), "c h w -> h w c").numpy().copy()
         lbl = labels[b].detach().cpu().numpy()
-        mat = cov_mat[b].detach().cpu().numpy()
+        mat = cov_mat[b].detach().cpu().float().numpy()
 
         rows_sub = np.arange(step // 2, H, step)
         cols_sub = np.arange(step // 2, W, step)
@@ -256,9 +256,9 @@ def _render_dir_quiver(
 
     result = []
     for b in range(B):
-        bg = rearrange(img_rgb[b].detach().cpu(), "c h w -> h w c").numpy().copy()
+        bg = rearrange(img_rgb[b].detach().cpu().float(), "c h w -> h w c").numpy().copy()
         lbl = labels[b].detach().cpu().numpy()
-        d = dir_val[b].detach().cpu().numpy()
+        d = dir_val[b].detach().cpu().float().numpy()
 
         # Channel layout: 2D → [x, y];  3D → [z, y, x].
         # For 2D quiver display we need (U=horizontal, V=vertical).
